ref code: do not clear the secnonce in deterministic_sign
Clearing the secnonce doesn't serve a purpose because it's a local variable. On
the other hand, we clear the secnonce in test_sign_and_verify_random to properly
demonstrate to users how to sign. Nowhere in the code do we try to erase
temporary secrets (such as k_1, k_2) from memory.